requirements.txt
googlegenai
streamlit